# backend/docs/specs/openapi.yml
openapi: 3.0.3
info:
  title: Sabor API
  description: API para el sistema de restaurante Sabor - Backend con autenticación JWT, gestión de productos, pedidos, reservas y usuarios
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: dev@sabor.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Servidor de desarrollo local
  - url: https://api.sabor.com
    description: Servidor de producción
paths:
...
  /api/reservas/horarios-disponibles:
    get:
      tags:
        - Reservas
      summary: Obtener horarios disponibles
      description: Retorna los horarios disponibles para reservas
      parameters:
        - name: fecha
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Fecha para verificar disponibilidad
      responses:
        '200':
          description: Lista de horarios disponibles
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: "12:00"
        '400':
          description: Error en la validación de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reservas/disponibilidad:
    get:
      tags:
        - Reservas
      summary: Verificar disponibilidad
      description: Verifica si hay disponibilidad para una fecha y hora específica
      parameters:
        - name: fecha
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Fecha de la reserva
        - name: hora
          in: query
          required: true
          schema:
            type: string
            pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
          description: Hora de la reserva (formato HH:MM)
      responses:
        '200':
          description: Información de disponibilidad
          content:
            application/json:
              schema:
                type: object
                properties:
                  disponible:
                    type: boolean
                    example: true
                  mensaje:
                    type: string
                    example: "Horario disponible"
        '400':
          description: Error en la validación de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reservas/hacerReserva:
    post:
      tags:
        - Reservas
      summary: Crear nueva reserva
      description: Crea una nueva reserva en el sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - personas
                - fecha
                - hora
                - nombre
                - telefono
                - email
              properties:
                personas:
                  type: integer
                  minimum: 1
                  example: 4
                fecha:
                  type: string
                  format: date
                  example: "2024-01-15"
                hora:
                  type: string
                  pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
                  example: "19:00"
                nombre:
                  type: string
                  example: "Juan Pérez"
                telefono:
                  type: string
                  pattern: ^\d{10}$
                  example: "1234567890"
                email:
                  type: string
                  format: email
                  example: "juan@example.com"
                peticiones:
                  type: string
                  example: "Mesa cerca de la ventana"
      responses:
        '201':
          description: Reserva creada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Reserva creada con éxito!"
                  reservaId:
                    type: integer
                    example: 123
                  fecha:
                    type: string
                    format: date
                    example: "2024-01-15"
                  hora:
                    type: string
                    example: "19:00"
        '400':
          description: Error en la validación de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Cuenta no verificada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reservas/historial:
    get:
      tags:
        - Reservas
      summary: Obtener historial de reservas
      description: Retorna el historial de reservas del usuario autenticado
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Historial de reservas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reserva'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reservas:
    get:
      tags:
        - Reservas
      summary: Obtener todas las reservas (Admin)
      description: Retorna todas las reservas del sistema (solo para administradores)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de todas las reservas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reserva'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Sin permisos suficientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: Función no implementada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reservas/{id}:
    get:
      tags:
        - Reservas
      summary: Obtener reserva por ID (Admin)
      description: Retorna una reserva específica por su ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID de la reserva
      responses:
        '200':
          description: Reserva encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reserva'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Sin permisos suficientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reserva no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: Función no implementada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Reservas
      summary: Actualizar reserva (Admin)
      description: Actualiza una reserva existente
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID de la reserva
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                personas:
                  type: integer
                  minimum: 1
                fecha:
                  type: string
                  format: date
                hora:
                  type: string
                  pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
                notas:
                  type: string
      responses:
        '200':
          description: Reserva actualizada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Reserva actualizada exitosamente
        '400':
          description: Error en la validación de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Sin permisos suficientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reserva no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: Función no implementada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Reservas
      summary: Eliminar reserva (Admin)
      description: Elimina una reserva del sistema
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID de la reserva
      responses:
        '200':
          description: Reserva eliminada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Reserva eliminada exitosamente
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Sin permisos suficientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Reserva no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: Función no implementada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # =================
  # HORARIOS ENDPOINTS
  # =================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido mediante login
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Error message"
        error:
          type: string
          example: "Detailed error information"
        errors:
          type: object
          additionalProperties:
            type: string
      required:
        - message
...
    Reserva:
      type: object
      properties:
        id_reservacion:
          type: integer
          example: 123
        id_cliente:
          type: integer
          example: 456
        numero_personas:
          type: integer
          example: 4
        fecha_reservacion:
          type: string
          format: date
          example: "2024-01-15"
        hora_reservacion:
          type: string
          pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
          example: "19:00"
        notas:
          type: string
          example: "Mesa cerca de la ventana"
        id_mesa:
          type: integer
          example: 5
        id_estado:
          type: integer
          example: 1
        fecha_creacion:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
...